Enable writing raw Rprof output to user-defined file
Added an input argument "prof_path" to the function "lineprof" that can be used to specify the path
where the raw Rprof output should be saved. The default value ("") causes the program to create a
temporary file for Rprof output and delete it upon exit - the same behavior as the original program.